win32: less warnings

+++ src/adt/Allocator.hh
@@ -20,9 +20,9 @@ constexpr size_t SIZE_1G = SIZE_1M * SIZE_1K;
 /* Base allocator interface */
 struct Allocator
 {
-    virtual void* alloc(size_t memberCount, size_t memberSize) = 0;
+    virtual void* alloc(u32 memberCount, u32 memberSize) = 0;
     virtual void free(void* p) = 0;
-    virtual void* realloc(void* p, size_t size) = 0;
+    virtual void* realloc(void* p, u32 size) = 0;
 };
 
 } /* namespace adt */
